Reduced memory usage.

+++ globals.c
@@ -3,7 +3,7 @@
 float DELTA_TICKS = 0;
 
 GLuint BOUND_TEXTURE = 0;
-int BLOCK_SIZE = 48;
+uint8_t BLOCK_SIZE = 48;
 
 GLfloat camera_x = 0.0f, camera_y = 0.0f;
 GLfloat mouse_x = 0.0f, mouse_y = 0.0f;
